# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Innovation Circle API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/request-link:
    post:
      operationId: requestMagicLink
      tags: [auth]
      summary: Request a magic login link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MagicLinkRequest"
      responses:
        "200":
          description: Magic link sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid email address
        "502":
          description: Email delivery failed
        "503":
          description: Email delivery or public URL is not configured
  /auth/verify:
    post:
      operationId: verifyMagicLink
      tags: [auth]
      summary: Verify a magic link token and create session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MagicLinkVerify"
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendee"
        "400":
          description: Invalid, expired, or already-used token
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendee"
        "401":
          description: Not authenticated

  # Circles
...
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List goals (own goals for attendee, all for admin)
      parameters:
        - name: attendeeId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: showCompleted
          in: query
          required: false
          schema:
            type: boolean
        - name: circleId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GoalWithAttendee"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/summary:
    get:
      operationId: getGoalsSummary
      tags: [goals]
      summary: Get goals summary (counts by status)
      parameters:
        - name: circleId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Goals summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalsSummary"
  /goals/{id}:
    get:
      operationId: getGoal
      tags: [goals]
      summary: Get a goal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "404":
          description: Not found
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Surveys
  /surveys:
    get:
      operationId: listSurveys
      tags: [surveys]
      summary: List surveys
      responses:
        "200":
          description: List of surveys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Survey"
    post:
      operationId: createSurvey
      tags: [surveys]
      summary: Create a survey (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SurveyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
  /surveys/{id}:
    get:
      operationId: getSurvey
      tags: [surveys]
      summary: Get a survey by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Survey
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
        "404":
          description: Not found
  /surveys/{id}/responses:
    post:
      operationId: submitSurveyResponse
      tags: [surveys]
      summary: Submit a survey response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SurveyResponseInput"
      responses:
        "201":
          description: Response submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurveyResponse"
    get:
      operationId: getSurveyResponses
      tags: [surveys]
      summary: Get responses for a survey (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Survey responses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SurveyResponse"

  # Suggestions
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # Auth
    MagicLinkRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    MagicLinkVerify:
      type: object
      required: [token]
      properties:
        token:
          type: string
          minLength: 1

    # Circle
...
    Attendee:
      type: object
      required: [id, name, email, company, role, circleId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        company:
          type: string
        role:
          type: string
        circleId:
          type: integer
        createdAt:
          type: string
...
    Goal:
      type: object
      required: [id, attendeeId, timeframe, status, createdAt]
      properties:
        id:
          type: integer
        attendeeId:
          type: integer
        timeframe:
          type: string
        status:
          type: string
        comments:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    GoalWithAttendee:
      type: object
      required: [id, attendeeId, timeframe, status, createdAt]
      properties:
        id:
          type: integer
        attendeeId:
          type: integer
        timeframe:
          type: string
        status:
          type: string
        comments:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        attendeeName:
          type: string
        attendeeCompany:
          type: string
    GoalInput:
      type: object
      required: [timeframe, status]
      properties:
        timeframe:
          type: string
        status:
          type: string
        comments:
          type: string
        dueDate:
          type: ["string", "null"]
    GoalUpdate:
      type: object
      properties:
        timeframe:
          type: string
        status:
          type: string
        comments:
          type: string
        dueDate:
          type: ["string", "null"]
    GoalsSummary:
      type: object
      required: [total, byStatus]
      properties:
        total:
          type: integer
        byStatus:
          type: object
          properties:
            new:
              type: integer
            inProgress:
              type: integer
            completed:
              type: integer
            notStarted:
              type: integer

    # Survey
    Survey:
      type: object
      required: [id, meetingId, type, questions, createdAt]
      properties:
        id:
          type: integer
        meetingId:
          type: integer
        type:
          type: string
        questions:
          type: array
          items:
            type: string
        createdAt:
          type: string
    SurveyInput:
      type: object
      required: [meetingId, type, questions]
      properties:
        meetingId:
          type: integer
        type:
          type: string
        questions:
          type: array
          items:
            type: string
    SurveyResponse:
      type: object
      required: [id, surveyId, attendeeId, answers, createdAt]
      properties:
        id:
          type: integer
        surveyId:
          type: integer
        attendeeId:
          type: integer
        answers:
          type: array
          items:
            type: string
        createdAt:
          type: string
    SurveyResponseInput:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            type: string

    # Suggestion
...