# lib/api-zod/src/generated/types/magicLinkVerify.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Innovation Circle API
 * OpenAPI spec version: 0.1.0
 */

export interface MagicLinkVerify {
  /** @minLength 1 */
  token: string;
}